feat: Support all autosquash prefixes
Committed, until now, only supported the `fixup!` prefix. The standard
Git implementation of `--autosquash` also supports `squash!` and
`amend!`. This commit modifies the logic of the handling of `fixup!` to
support the other two prefixes in the same way.

Note that this purely changes the logic. Function names and options
remain unchanged to ensure backwards compatibility. It may be worth
considering renaming this to `autosquash` given it is motivated
primarily by Git's `--autosquash` flag.

Ref: https://git-scm.com/docs/git-rebase#Documentation/git-rebase.txt---autosquash
Signed-off-by: JP-Ellis <josh@jpellis.me>